feat: refactor voting types and integrate extended vote information across components
